refactor(raijin): own project model
