+++ yarn/plugin-checks/package.json
@@ -25,6 +25,7 @@
     "@atls/code-lint": "workspace:*",
     "@atls/code-test": "workspace:*",
     "@atls/code-typescript": "workspace:*",
+    "@atls/raijin": "workspace:*",
     "@atls/yarn-plugin-files": "workspace:*",
     "@atls/yarn-plugin-tools": "workspace:*",
     "@atls/yarn-plugin-workspaces": "workspace:*",
@@ -34,7 +35,6 @@
     "strip-ansi": "7.1.0"
   },
   "devDependencies": {
-    "@atls/raijin": "workspace:*",
     "@octokit/types": "13.5.0",
     "@types/babel__code-frame": "7.0.6",
     "@types/node": "24.12.2",
