+++ yarn/plugin-tools/sources/index.ts
@@ -1,2 +1,1 @@
 export { plugin as default } from './tools.plugin.js'
-export *                     from './project/index.js'
