+++ yarn/plugin-typescript/sources/typecheck.command.tsx
@@ -21,10 +21,10 @@ import { TypeScriptProgress }             from '@atls/cli-ui-typescript-progress
 import { TypeScript }                     from '@atls/code-typescript'
 import { COMMAND_PROXY_EXECUTION }        from '@atls/yarn-plugin-tools/command-context'
 import { renderStatic }                   from '@atls/cli-ui-renderer-static-component'
+import { createProjectModel }             from '@atls/raijin/project'
 import { createCommandProxyEnvironment }  from '@atls/yarn-plugin-tools/command-context'
 import { resolveWorkspaceCommandContext } from '@atls/yarn-plugin-tools/command-context'
 import { makeCurrentYarnExecutable }      from '@atls/yarn-plugin-tools/current-yarn-executable'
-import { createProjectModel }             from '@atls/yarn-plugin-tools/project'
 
 export class TypeCheckCommand extends BaseCommand {
   static override paths = [['typecheck']]
