+++ yarn/raijin/src/project/manifest.ts
